move MaskHackClient from mask package to tile package
