+++ render-ws-spark-client/src/main/java/org/janelia/render/client/spark/pipeline/AlignmentPipelineStepId.java
@@ -4,7 +4,7 @@ import java.util.function.Supplier;
 
 import org.janelia.render.client.spark.MipmapClient;
 import org.janelia.render.client.spark.ScapeClient;
-import org.janelia.render.client.spark.mask.MaskHackClient;
+import org.janelia.render.client.spark.tile.MaskHackClient;
 import org.janelia.render.client.spark.match.ClusterCountClient;
 import org.janelia.render.client.spark.match.CopyMatchClient;
 import org.janelia.render.client.spark.match.MultiStagePointMatchClient;
